requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
PyYAML==6.0
requests==2.28.2
urllib3==1.26.14
